Try to resize the iframe based on content, scrollers only when needed
Uses iframeResizer library that updates the size of the iframe element based
on the changes to the size of the content contained in the iframe (the iframe
must include iframeResizerContentwindow script for this to work).

Also add a static sized div inside the fixed dialog curtain that adds scrollers
when needed.