Add a way of showing an admin page wrapping main content in an iframe
Implement a firts attempt for showing admin-page content inside an iframe. The index page to be used for the admin
page can contain Atlassian decorators in addition to the content that will be embedded inside an iframe. Depending
on the query parameters, the decorator content or the main app (body, scripts...) content will be stripped. When
the page is loaded with query parameter telling that it will be the iframe content, all the Atlassian decorators are
stripped, otherwise the main page content is stripped and instead an iframe will be placed that will have as its source
the same page but with the query parameter telling to load it as the iframe content.