Insert the iframeResizer.contentWindow to content loaded for iframe automatically
When the servlet is called with the parameter telling that the file is meant to be loaded to be content of a iframe,
the servlet adds a script element to the loaded document containing the contentWindow side of the iframeResizer
library. In this way the plugin developer using SPARK don't have to do anything to get the iframeResizer functionality.