Convert css to js that can be concatenated to the dist JS-file in gulp build
Use a gulp-plugin that takes a JS-file and turns it into a JS-script that will add a style element to the head
with the styles from the CSS-file. This JS-script can then be concatenated to the main JS dist-file, so that
a plugin-developer is not required to add a separate download-resource for the CSS-file.