Add a gulp build task that prepares files for 'dist'
Add a task that cleans old files from 'dist' directory, runs tests, builds the needed files, and then copies
the resulting built JS dist-file, and other needed files into the 'dist' directory. The source files are copied
also such (in addition to concatenating them to the JS dist-file) to support old SPARK style of including
the files individually. The iframeResizer-files are also copied also as such because they are needed also
for the backend SPARK iframe operation modes.

This task is meant as an easy entry point for Maven to run the build and instruct gulp to copy the 'dist'
files to the correct place in the 'target' directory.