Add a simple no-conflict / version handling
Restore the old SPARK object after the new one is initialized and stored in a special SPARK.__versions object. Make
it possible to get latest or a specific version of SPARK by calling SPARK.__versions.get() method.

It is expected that the bundled iframeresizer-script can handle possible version conflicts itself and that there is
also relatively good interoperability in the possible case that the host and contentWindow don't use exactly
same version of iframeResizer.