Fix a bug with usage of soy templates in different context
To get a soy output that matches the expected output 'autoescape="deprecated-contextual"' has to be used with
the soynode compiler. When the template is included directly as a webresource though, the autoescape attribute
cannot be used because it causes a problem with the soy compiler used there. Thus the needed argument is now
added as a build step in the gulpfile (and the source file that will be also copied into the result is left without
the attribute).