First version of improved context passing
Adds the context data to be included in the iframe as an extra property
on the iframe DOM element. This can be accessed from the iframe context
as frameElement (as long as the domain is same).

The access to the frameElement is abstracted by adding a script to the
contentWindow wrapping the logic (which also makes it possible to refactor
how the data exactly is passed in the future if needed).

This implementation removes race conditions (the data is always ready when
the iframe content can be loaded), and also ensures that the context data
is available also if the contentWindow of the iframe changes (eg. by refreshing
the embedded document).