Add a few tests checking that the actions and templates they use are in sync
The tests test that the template (or example template) meant to be used with certain action does not use
any variables that are not present in the action or known to be in the Velocity context when rendering
because of other reasons.

The tests don't add any value over eg. IntelliJ's capabilities as such, but because they are integrated in
the build, it is harder to ignore them