DEV-109: Use require for loading AJS / jQuery only if globals not available
The change is needed because window.require('ajs') does not work on Jira, at least not yet.
On Confluence too the main motivation for using AMD module loading was that it has been announced
that the globals will be deprecated / removed: There should be no harm in actually using the
globals if they are still defined on Confluence either.